Held karp algorithm implemented problems

+++ src/City.java
@@ -32,8 +32,8 @@ public class City {
 
     // Gets the distance to given city
     public double distanceTo(City city){
-        int xDistance = Math.abs(getX() - city.getX());
-        int yDistance = Math.abs(getY() - city.getY());
+        double xDistance = Math.abs(getX() - city.getX());
+        double yDistance = Math.abs(getY() - city.getY());
         double distance = Math.sqrt( (xDistance*xDistance) + (yDistance*yDistance) );
 
         return distance;
